[#91995] modelwrappers: anomaly_detection: cnn: Converting data to float32 (as specified in iospec) in _preprocess_input()

+++ kenning/modelwrappers/anomaly_detection/cnn.py
@@ -274,7 +274,7 @@ class PyTorchAnomalyDetectionCNN(PyTorchWrapper, AutoPyTorchModel):
     def preprocess_input(self, X) -> List[Any]:
         X = np.asarray(X[0])
         X = (X - self.mean) / (3 * self.std)
-        X = np.expand_dims(X, -3)
+        X = np.float32(np.expand_dims(X, -3))
         return [X]
 
     def postprocess_outputs(self, y: List[Any]) -> List[np.ndarray]:
